+++ src/components/ModalOverlay/types.ts
@@ -1,8 +1,4 @@
-import type { ReactChild } from 'react';
-
 export type Props = {
-  children: ReactChild;
   onClick: () => void;
-  closeEscape?: boolean;
   className?: string;
 };
